% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{format_risk_level}
\alias{format_risk_level}
\title{Format Risk Level}
\usage{
format_risk_level(ffdi)
}
\arguments{
\item{ffdi}{Numeric vector of FFDI values}
}
\value{
Character vector of risk levels
}
\description{
Converts numeric FFDI values to categorical risk levels.
}
\details{
Risk levels are defined as:
\itemize{
  \item Low: FFDI < 12
  \item Moderate: 12 <= FFDI < 24
  \item High: 24 <= FFDI < 50
  \item Very High: 50 <= FFDI < 75
  \item Extreme: FFDI >= 75
}
}
\examples{
format_risk_level(c(8, 15, 30, 60, 80))

}
